pyproject.toml
[tool.poetry]
name = "flowtrail"
version = "0.0.1"
description = "trail for redis"
authors = ["anantsrivastava02  <itanart333@gmail.com>"]
license = "0.0.099"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.8.0"
redis = "^5.2.0"
pandas = "^2.1.1"
beautifulsoup4 = "^4.12.2"
mysql-connector-python = "^9.1.0"
selenium = "^4.26.1"
mysql-connector = "^2.2.9"
python-dotenv = "^1.0.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
